requirements.txt
obs-websocket-py>=1.0
python-dotenv>=1.0.0
